Implement proper timing
Before, we almost never delayed because we were aiming for 100Hz which is also the tick frequency.
So after executing one step, we already had less than one tick left to delay, so vTaskDelay did not delay at all.

xTaskDelayUntil could actually handle 100Hz because it just waits for the next tick, even if there is less than one tick left.
But 50Hz is plenty; if it is good enough for videos, it is good enough for my LEDs.